requirement.txt
asgiref==3.8.1
Django==4.2
djangorestframework==3.15.1
sqlparse==0.4.4
gunicorn==21.2.0
